[CoreBundle] allow purchasable as gifts
