+++ src/CoreShop/Component/Core/Cart/Rule/Action/GiftProductActionProcessor.php
@@ -18,6 +18,7 @@ declare(strict_types=1);
 
 namespace CoreShop\Component\Core\Cart\Rule\Action;
 
+use CoreShop\Bundle\ResourceBundle\Pimcore\Repository\StackRepository;
 use CoreShop\Component\Order\Cart\Rule\Action\CartPriceRuleActionProcessorInterface;
 use CoreShop\Component\Order\Factory\AdjustmentFactoryInterface;
 use CoreShop\Component\Order\Factory\OrderItemFactoryInterface;
@@ -26,13 +27,12 @@ use CoreShop\Component\Order\Model\OrderInterface;
 use CoreShop\Component\Order\Model\OrderItemInterface;
 use CoreShop\Component\Order\Model\PriceRuleItemInterface;
 use CoreShop\Component\Order\Model\PurchasableInterface;
-use CoreShop\Component\Product\Repository\ProductRepositoryInterface;
 use CoreShop\Component\Rule\Model\ActionInterface;
 
 final class GiftProductActionProcessor implements CartPriceRuleActionProcessorInterface
 {
     public function __construct(
-        private ProductRepositoryInterface $productRepository,
+        private StackRepository $productRepository,
         private OrderItemFactoryInterface $cartItemFactory,
         private AdjustmentFactoryInterface $adjustmentFactory,
     ) {
@@ -67,6 +67,7 @@ final class GiftProductActionProcessor implements CartPriceRuleActionProcessorIn
         }
 
         $item = $this->cartItemFactory->createWithCart($cart, $product);
+        $item->setQuantity(1);
         $item->setIsGiftItem(true);
 
         $adjustment = $this->adjustmentFactory->createWithData(
